Move tests back to __test__ directory
