+++ package.json
@@ -17,11 +17,11 @@
         "test:debug": "cross-env CORE_ENV=test node --inspect-brk ./node_modules/.bin/jest --runInBand",
         "test:watch": "cross-env CORE_ENV=test jest --runInBand --watch",
         "test:watch:all": "cross-env CORE_ENV=test jest --runInBand --watchAll",
-        "test:unit": "cross-env jest __tests__/unit/** packages/*/src/** --forceExit",
-        "test:unit:coverage": "cross-env jest __tests__/unit packages/*/src/** --coverage --coveragePathIgnorePatterns='/(defaults.ts|index.ts)$'",
-        "test:unit:debug": "cross-env node --inspect-brk ./node_modules/.bin/jest __tests__/unit packages/*/src/**",
-        "test:unit:watch": "cross-env jest __tests__/unit packages/*/src/** --watch",
-        "test:unit:watch:all": "cross-env CORE_ENV=test jest __tests__/unit packages/*/src/** --watchAll",
+        "test:unit": "cross-env jest __tests__/unit/** --forceExit",
+        "test:unit:coverage": "cross-env jest __tests__/unit --coverage --coveragePathIgnorePatterns='/(defaults.ts|index.ts)$'",
+        "test:unit:debug": "cross-env node --inspect-brk ./node_modules/.bin/jest __tests__/unit",
+        "test:unit:watch": "cross-env jest __tests__/unit --watch",
+        "test:unit:watch:all": "cross-env CORE_ENV=test jest __tests__/unit --watchAll",
         "test:integration": "cross-env CORE_ENV=test jest __tests__/integration --runInBand --forceExit",
         "test:integration:coverage": "cross-env CORE_ENV=test jest __tests__/integration --coverage --coveragePathIgnorePatterns='/(defaults.ts|index.ts)$' --runInBand --forceExit",
         "test:integration:debug": "cross-env CORE_ENV=test node --inspect-brk ./node_modules/.bin/jest __tests__/integration --runInBand",
